SubSyncRosa
I half-built this utility in REALBasic some ten years ago to synchronize SRT subtitles and left it half-way since I didn't need more functionality.

Finally putting here as-is, in case anybody can take something good from it. It's built in a rush on top of a separate application, so it may have a lot of extraneous code and modules.

Doesn't use the standard listbox for some reason I can't recall.

Usage: Open one SRT, tell it you want to sync it another, select start and endpoints and let it calculate all the adjusted timings.

Obvs won't work for subtitles for different versions of a movie (added/deleted/different scenes, etc.) but as a way to read and write SRTs it may be useful to somebody.